requirements.txt
pillow
pytest
pytesseract
numpy
opencv-python
pywavelets
fastapi
uvicorn[standard]
